fix: unit tests with refactor
