+++ spid_cie_oidc/authority/tests/test_02_trust_anchor_intermediary.py
@@ -1,0 +1,65 @@
+from django.test import TestCase, Client, override_settings
+from django.urls import reverse
+
+from spid_cie_oidc.entity.jwtse import verify_jws
+from spid_cie_oidc.entity.models import *
+from spid_cie_oidc.entity.trust_chain import trust_chain_builder
+from spid_cie_oidc.entity.statements import EntityConfiguration, get_entity_configurations
+from spid_cie_oidc.entity.tests.settings import *
+
+from spid_cie_oidc.authority.models import *
+from spid_cie_oidc.authority.tests.settings import *
+
+from unittest.mock import patch
+
+import datetime
+
+
+class TATest(TestCase):
+
+    def setUp(self):
+        self.ta_conf = FederationEntityConfiguration.objects.create(**ta_conf_data)
+        self.rp_conf = FederationEntityConfiguration.objects.create(**rp_conf)
+
+        self.rp_profile = FederationEntityProfile.objects.create(**RP_PROFILE)
+        self.rp = FederationDescendant.objects.create(**rp_onboarding_data)
+
+        self.rp_jwk = PublicJwk.objects.create(
+            jwk = self.rp_conf.public_jwks[0],
+            kid = self.rp_conf.public_jwks[0]['kid']
+        )
+        FederationDescendantJwk.objects.create(
+            descendant=self.rp,
+            jwk = self.rp_jwk
+        )
+        FederationEntityAssignedProfile.objects.create(
+            descendant = self.rp,
+            profile = self.rp_profile,
+            issuer = self.ta_conf
+        )
+
+    def test_fetch_endpoint(self):
+        url = reverse('oidcfed_fetch')
+        c = Client()
+        res = c.get(url, data={'sub': self.rp.sub})
+        data = verify_jws(res.content.decode(), self.ta_conf.jwks[0])
+        self.assertTrue(
+            data['jwks']
+        )
+
+    def test_list_endpoint(self):
+        url = reverse('oidcfed_list')
+        c = Client()
+        res = c.get(url, data={'is_leaf': True})
+        self.assertTrue(
+            res.json()[0] == self.rp.sub
+        )
+        self.assertEqual(res.status_code, 200)
+        
+        res = c.get(url, data={'is_leaf': False})
+        self.assertFalse(res.json())
+        self.assertEqual(res.status_code, 200)
+        
+        res = c.get(url, data={})
+        self.assertTrue(res.json())
+        self.assertEqual(res.status_code, 200)
