+++ spid_cie_oidc/onboarding/views.py
@@ -1,3 +1,4 @@
+from django.core.paginator import Paginator
 from django.shortcuts import render
 
 from .forms import OnboardingRegistrationForm
